Fix ALL the errors and uncompleted functions occured when testing until it succeed to run peers and orderers on target servers. In test, this program can build & run a fabric network (only system channel)

+++ run.sh
@@ -1,14 +1,20 @@
 #! /bin/sh
-python3 main.py -c initnode -C config.yaml -U config.new.yaml >> output.temp
+if [ "$1" = "-h" ]
+then
+    echo 'run.sh <command> <config_file> <updated_config_file>'
+else
+    rm output.temp
+    python3 main.py -c $1 -C $2 -U $3 >> output.temp
 
-cat output.temp | while read line
-do
-    # echo $line >> 'output'
-    echo ------------------------------
-    echo $line
-    echo ------------------------------
-    `$line`
-done
+    cat output.temp | while read line
+    do
+        # echo $line >> 'output'
+        echo '>>>>>--------------------------------------------------'
+        echo $line
+        $line
+        echo '<<<<<--------------------------------------------------'
+    done
+    rm output.temp
+fi
 
-rm output.temp
